feat: group new topic page (WIP)
