+++ packages/design/components/EditorForm/style/index.less
@@ -7,6 +7,7 @@
   &__container {
     display: inline-flex;
     flex-direction: column;
+    align-self: stretch;
     min-height: 162px;
     border: 2px solid @gray-10;
     border-radius: 12px;
@@ -32,7 +33,6 @@
     flex: 1 0 auto;
     padding: 0;
     min-width: @toolbox-width;
-    width: @text-width;
     border: none;
     outline: none;
     color: @gray-100;
