+++ packages/website/src/pages/index/group/[name]/index/forum.tsx
@@ -7,7 +7,7 @@ import { useGroupRecentTopics } from '@bangumi/website/hooks/use-group';
 import { useTransitionNavigate } from '@bangumi/website/hooks/use-navigate';
 import { usePaginationParams } from '@bangumi/website/hooks/use-pagination';
 
-import TopicsTable from './components/TopicsTable';
+import TopicsTable from '../components/TopicsTable';
 import styles from './style.module.less';
 
 const GroupForum = () => {
