+++ packages/website/src/pages/index/group/[name]/new_topic.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { Outlet } from 'react-router-dom';
+
+import ErrorBoundary from '@bangumi/website/components/ErrorBoundary';
+
+const GroupNewTopicPage = () => {
+  return (
+    <ErrorBoundary fallback={<>Group Not found</>}>
+      <Outlet />
+    </ErrorBoundary>
+  );
+};
+
+export default GroupNewTopicPage;
